The new feature has been add: filters to search the users and pass it to frontend and to the future save algorithm

+++ source/new_user.c
@@ -2,12 +2,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 
 
 PTR_users START_ptr = NULL;
 
 
-static int create_user(PTR_users *INIT_ptr, const char *name, const int year){
+static int create_user(PTR_users *INIT_ptr, const char *name, const uint_least8_t year){
 
     
  PTR_users NEW_node = NULL;
@@ -34,7 +35,7 @@ static int create_user(PTR_users *INIT_ptr, const char *name, const int year){
 }
 
 
-int new_user(const char *name, const int year){
+int new_user(const char *name, const uint_least8_t year){
 
     switch (create_user(&START_ptr ,name, year)){
         case 0:
